Convert jsx files to tsx, add global styles, initial App.tsx layout
