+++ src/components/Footer/Footer.tsx
@@ -1,0 +1,7 @@
+import "./Footer.scss";
+
+function Footer() {
+  return <>This is the Footer.</>;
+}
+
+export default Footer;
