+++ src/components/Header/Header.tsx
@@ -1,0 +1,7 @@
+import "./Header.scss";
+
+function Header() {
+  return <>This is the Header.</>;
+}
+
+export default Header;
