--- src/pages/HomePage/HomePage.jsx
@@ -1,7 +1,0 @@
-import "../HomePage/HomePage.scss";
-
-function HomePage() {
-  return <></>;
-}
-
-export default HomePage;
