+++ src/pages/HomePage/HomePage.tsx
@@ -1,0 +1,7 @@
+import "./HomePage.scss";
+
+function HomePage() {
+  return <>This is the Home Page.</>;
+}
+
+export default HomePage;
