+++ src/styles/partials/_global.scss
@@ -1,0 +1,10 @@
+@use "./colors" as *;
+@use "./fonts" as *;
+@use "./mixins" as *;
+@use "./variables" as *;
+* {
+  padding: 0;
+  margin: 0;
+  list-style-type: none;
+  text-decoration: none;
+}
