+++ src/styles/partials/_mixins.scss
@@ -1,0 +1,3 @@
+@use "./colors" as *;
+@use "./fonts" as *;
+@use "./variables" as *;
